requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
executing==1.2.0
fonttools==4.39.4
idna==3.4
kiwisolver==1.4.4
matplotlib-inline==0.1.6
numpy==1.24.3
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycparser==2.21
PyGithub==1.58.2
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyX==0.16
requests==2.31.0
scapy==2.5.0
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
wrapt==1.15.0
